requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.1
pytz==2023.3.post1
gunicorn==21.2.0
pymysql==1.1.0
cryptography==42.0.5
